requirements.txt
dlib==19.24.0
face-recognition==1.3.0
numpy==1.24.2
Pillow==10.2.0
pyside6==6.6.1
opencv-python==4.9.0.80
